Split transmission tasks
