+++ flexget/plugins/clients/transmission/transmission.py
@@ -1,24 +1,20 @@
 from __future__ import unicode_literals, division, absolute_import
 from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from past.builtins import basestring
-from future.moves.urllib.parse import urlparse
 from future.utils import text_to_native_str
 
 import os
 import logging
 import base64
-import re
 import time
 from datetime import datetime
 from datetime import timedelta
 from netrc import netrc, NetrcParseError
 
-from flexget import plugin, validator
-from flexget.entry import Entry
+from flexget import plugin
 from flexget.event import event
 from flexget.utils.template import RenderError
 from flexget.utils.pathscrub import pathscrub
-from flexget.utils.tools import parse_timedelta
 
 from flexget.config_schema import one_or_more
 from fnmatch import fnmatch
@@ -35,7 +31,6 @@ log = logging.getLogger('transmission')
 
 
 class TransmissionBase(object):
-
     def __init__(self):
         self.client = None
         self.opener = None
@@ -143,61 +138,6 @@ class TransmissionBase(object):
                     log.info('Successfully connected to transmission.')
                 else:
                     log.error('It looks like there was a problem connecting to transmission.')
-
-
-class PluginTransmissionInput(TransmissionBase):
-
-    def validator(self):
-        """Return config validator"""
-        root = validator.factory()
-        root.accept('boolean')
-        advanced = root.accept('dict')
-        self._validator(advanced)
-        advanced.accept('boolean', key='onlycomplete')
-        return root
-
-    def prepare_config(self, config):
-        config = TransmissionBase.prepare_config(self, config)
-        config.setdefault('onlycomplete', True)
-        return config
-
-    def on_task_input(self, task, config):
-        config = self.prepare_config(config)
-        if not config['enabled']:
-            return
-
-        if not self.client:
-            self.client = self.create_rpc_client(config)
-        entries = []
-
-        # Hack/Workaround for http://flexget.com/ticket/2002
-        # TODO: Proper fix
-        if 'username' in config and 'password' in config:
-            self.client.http_handler.set_authentication(self.client.url, config['username'], config['password'])
-
-        session = self.client.get_session()
-
-        for torrent in self.client.get_torrents():
-            downloaded, bigfella = self.torrent_info(torrent, config)
-            seed_ratio_ok, idle_limit_ok = self.check_seed_limits(torrent, session)
-            if not config['onlycomplete'] or (downloaded and
-                                              ((
-                                                  torrent.status == 'stopped' and
-                                                  seed_ratio_ok is None and
-                                                  idle_limit_ok is None) or
-                                               (seed_ratio_ok is True or idle_limit_ok is True))):
-                entry = Entry(title=torrent.name,
-                              url='file://%s' % torrent.torrentFile,
-                              torrent_info_hash=torrent.hashString,
-                              content_size=torrent.totalSize / (1024 * 1024))
-                for attr in ['comment', 'downloadDir', 'isFinished', 'isPrivate']:
-                    entry['transmission_' + attr] = getattr(torrent, attr)
-                entry['transmission_trackers'] = [t['announce'] for t in torrent.trackers]
-                # bigfella? Is this actually the path to the torrent file? see GitHub #1403
-                if bigfella:
-                    entry['location'] = bigfella
-                entries.append(entry)
-        return entries
 
 
 class PluginTransmission(TransmissionBase):
@@ -456,9 +396,9 @@ class PluginTransmission(TransmissionBase):
                     fl = cli.get_files(r.id)
 
                     if ('magnetization_timeout' in options['post'] and
-                        options['post']['magnetization_timeout'] > 0 and
+                                options['post']['magnetization_timeout'] > 0 and
                             not downloaded and
-                            len(fl[r.id]) == 0):
+                                len(fl[r.id]) == 0):
                         log.debug('Waiting %d seconds for "%s" to magnetize', options['post']['magnetization_timeout'],
                                   entry['title'])
                         fl = _wait_for_files(cli, r, options['post']['magnetization_timeout'])
@@ -514,7 +454,7 @@ class PluginTransmission(TransmissionBase):
                             dl_list.append(main_id)
                     elif find_main_file:
                         log.warning('No files in "%s" are > %d%% of content size, no files renamed.',
-                            entry['title'], main_ratio * 100)
+                                    entry['title'], main_ratio * 100)
 
                     # If we have a main file and want to rename it and associated files
                     if 'content_filename' in options['post'] and main_id is not None:
@@ -573,7 +513,7 @@ class PluginTransmission(TransmissionBase):
                 # if addpaused was defined and set to False start the torrent;
                 # prevents downloading data before we set what files we want
                 if ('paused' in options['post'] and not options['post']['paused'] or
-                        'paused' not in options['post'] and cli.get_session().start_added_torrents):
+                                'paused' not in options['post'] and cli.get_session().start_added_torrents):
                     cli.start_torrent(r.id)
                 elif options['post'].get('paused'):
                     log.debug('sleeping 5s to stop the torrent...')
@@ -598,107 +538,6 @@ class PluginTransmission(TransmissionBase):
     on_task_abort = on_task_learn
 
 
-class PluginTransmissionClean(TransmissionBase):
-    """
-    Remove completed torrents from Transmission.
-
-    Examples::
-
-      clean_transmission: yes  # ignore both time and ratio
-
-      clean_transmission:      # uses transmission's internal limits for idle time and seed ratio ( if defined )
-        transmission_seed_limits: yes
-
-      clean_transmission:      # matches time only
-        finished_for: 2 hours
-
-      clean_transmission:      # matches ratio only
-        min_ratio: 0.5
-
-      clean_transmission:      # matches time OR ratio
-        finished_for: 2 hours
-        min_ratio: 0.5
-
-    Default values for the config elements::
-
-      clean_transmission:
-        host: localhost
-        port: 9091
-        enabled: yes
-    """
-
-    def validator(self):
-        """Return config validator"""
-        root = validator.factory()
-        root.accept('boolean')
-        advanced = root.accept('dict')
-        self._validator(advanced)
-        advanced.accept('number', key='min_ratio')
-        advanced.accept('interval', key='finished_for')
-        advanced.accept('boolean', key='transmission_seed_limits')
-        advanced.accept('boolean', key='delete_files')
-        advanced.accept('regexp', key='tracker')
-        advanced.accept('regexp', key='preserve_tracker')
-        directories_re = advanced.accept('list', key='directories')
-        directories_re.accept('regexp')
-        return root
-
-    def on_task_exit(self, task, config):
-        config = self.prepare_config(config)
-        if not config['enabled'] or task.options.learn:
-            return
-        if not self.client:
-            self.client = self.create_rpc_client(config)
-        nrat = float(config['min_ratio']) if 'min_ratio' in config else None
-        nfor = parse_timedelta(config['finished_for']) if 'finished_for' in config else None
-        delete_files = bool(config['delete_files']) if 'delete_files' in config else False
-        trans_checks = bool(config['transmission_seed_limits']) if 'transmission_seed_limits' in config else False
-        tracker_re = re.compile(config['tracker'], re.IGNORECASE) if 'tracker' in config else None
-        preserve_tracker_re = re.compile(config['preserve_tracker'], re.IGNORECASE) if 'preserve_tracker' in config else None
-        directories_re = config.get('directories')
-
-        session = self.client.get_session()
-
-        remove_ids = []
-        for torrent in self.client.get_torrents():
-            log.verbose('Torrent "%s": status: "%s" - ratio: %s -  date added: %s - date done: %s' %
-                        (torrent.name, torrent.status, torrent.ratio, torrent.date_added, torrent.date_done))
-            downloaded, dummy = self.torrent_info(torrent, config)
-            seed_ratio_ok, idle_limit_ok = self.check_seed_limits(torrent, session)
-            tracker_hosts = (urlparse(tracker['announce']).hostname for tracker in torrent.trackers)
-            is_clean_all = nrat is None and nfor is None and trans_checks is False
-            is_minratio_reached = nrat and (nrat <= torrent.ratio)
-            is_transmission_seedlimit_unset = trans_checks and seed_ratio_ok is None and idle_limit_ok is None
-            is_transmission_seedlimit_reached = trans_checks and seed_ratio_ok is True
-            is_transmission_idlelimit_reached = trans_checks and idle_limit_ok is True
-            is_torrent_seed_only = torrent.date_done <= torrent.date_added
-            is_torrent_idlelimit_since_added_reached = nfor and (torrent.date_added + nfor) <= datetime.now()
-            is_torrent_idlelimit_since_finished_reached = nfor and (torrent.date_done + nfor) <= datetime.now()
-            is_tracker_matching = not tracker_re or any(tracker_re.search(host) for host in tracker_hosts)
-            is_preserve_tracker_matching = False
-            if preserve_tracker_re is not None:
-                is_preserve_tracker_matching = any(preserve_tracker_re.search(host) for host in tracker_hosts)
-            is_directories_matching = not directories_re or any(
-                re.compile(directory, re.IGNORECASE).search(torrent.downloadDir) for directory in directories_re)
-            if (downloaded and (is_clean_all or
-                                is_transmission_seedlimit_unset or
-                                is_transmission_seedlimit_reached or
-                                is_transmission_idlelimit_reached or
-                                is_minratio_reached or
-                                (is_torrent_seed_only and is_torrent_idlelimit_since_added_reached) or
-                                (not is_torrent_seed_only and is_torrent_idlelimit_since_finished_reached)) and
-                    is_directories_matching and (not is_preserve_tracker_matching and is_tracker_matching)):
-                if task.options.test:
-                    log.info('Would remove finished torrent `%s` from transmission', torrent.name)
-                    continue
-                log.info('Removing finished torrent `%s` from transmission', torrent.name)
-                remove_ids.append(torrent.id)
-        if remove_ids:
-            self.client.remove_torrent(remove_ids, delete_files)
-
-
 @event('plugin.register')
 def register_plugin():
     plugin.register(PluginTransmission, 'transmission', api_ver=2)
-    plugin.register(PluginTransmissionInput, 'from_transmission', api_ver=2)
-    plugin.register(PluginTransmissionClean, 'clean_transmission', api_ver=2)
